% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R PACKAGE (DEFINITIVO).R
\name{be_judgf}
\alias{be_judgf}
\title{Break-Even Result Evaluator}
\usage{
be_judgf(b)
}
\arguments{
\item{b}{A numeric value we obtained from sum_be().}
}
\value{
As logical, an evaluation on the numeric break-even value we have as argument of the function, following the UEFA regulations, according to which a deficit of more than €30million on a 3-years monitoring period cannot be accepted by the Commission.
}
\description{
The function prints as logical the evaluation of the numeric break-even results: 'SURPLUS' if positive, 'ACCEPTABLE' if falling within the limits established by the UEFA Commission, 'SANCTION' if negative exceeding €30million threshold.
}
