% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R PACKAGE (DEFINITIVO).R
\name{beyR}
\alias{beyR}
\title{Yearly Break-Even Function}
\usage{
beyR(df1, df2, x, y)
}
\arguments{
\item{df1}{The dataframe we obtained from rel_vaRR(), including the column obtained by running the trnsfR() function.}

\item{df2}{The dataframe we obtained from rel_vaRE(), including the column obtained by running the trnsfR() function.}

\item{x}{Number of the column with Aggregate Values in df1.}

\item{y}{Number of the column with Aggregate Values in df2.}
}
\value{
A vector, whose elements for yearly Break-even (difference by rows of two Aggregate Values columns)
}
\description{
The function creates a vector by applying a sum by row within different columns with aggregate values of interest (for break-even purposes) of different dataframes.
}
\examples{
z <- c(beyR(df1,df2,6,6))
}
