% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R PACKAGE (DEFINITIVO).R
\name{completefunc}
\alias{completefunc}
\title{Default Ordered Function}
\usage{
completefunc()
}
\value{
The function returns the evaluation as logical of the break-even result for the 3-years  monitoring period, together with the interactive bars graph.
}
\description{
The function runs the all the function implemented in the FFP.R package in the logical order.
}
