% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{ASR P/L Statement for a 3-years period 2016-2018.}
\format{A data frame with 3 rows and 24 variables (P/L statement accounts):
\describe{
  \item{Ricavi}{Revenues}
  \item{Diritti.Tv}{Broadcasting rights}
  \item{Proventi.gestione.diritti.calciatori}{Profit on players disposal}
  \item{Stadio}{Income from facilities management}
  \item{Attivita.commerciali}{Commercial activities}
  \item{Sponsorizzazioni}{Sponsorship}
  \item{Pubblicitari}{Marketing}
  \item{Altri.ricavi.e.proventi}{Other operating income}
  \item{Costi}{Costs}
  \item{Personale.tesserato}{Employees wages}
  \item{Oneri.gestione.diritti.calciatori}{Loss on disposal of players}
  \item{Spese.per.servizi}{Service charges}
  \item{Acquisti.di.materie.prime.e.di.consumo}{Cost of raw materials}
  \item{Spese.per.godimento.beni.di.terzi}{Loss in respect of assets retained by third parties}
  \item{Altro.personale}{Other employees expenses}
  \item{Oneri.diversi.di.gestione}{Other operating expenses}
  \item{EBITDA}{Earnings before income taxes,depreciation and amortization}
  \item{Amm.to.e.svalutaz.Diritti.calciatori}{Profit/Loss on players disposal due to amortization/impairment}
  \item{Amm.to.altre.attivita.materiali/immateriali}{Profit/Loss on disposal and amortization/impairment of certain tangible/intangible assets}
  \item{Svalutazione.crediti.correnti}{Write-down of current receivables}
  \item{Acc.to.fondi.rischi}{Other financial expenses}
  \item{EBIT}{Earnings before interests and taxes}
  \item{EBT}{Earnings before taxes}
  \item{Utile/Perdita.dell'esercizio}{Net profit/loss}
}}
\source{
\url{https://www.calcioefinanza.it/2018/07/19/bilancio-as-roma-2017-2018-preview-nainggolan-alisson/}
}
\usage{
data
}
\description{
This dataset contains the P/L statements of the ASR football club, for the financial years ending respectively 30-06-2016,30-06-2017,30-06-2018.
}
\keyword{datasets}
