% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R PACKAGE (DEFINITIVO).R
\name{frmtb}
\alias{frmtb}
\title{Suitable Format Function}
\usage{
frmtb(d, vct_date)
}
\arguments{
\item{d}{The dataframe of interest for the package.}

\item{vct_date}{A vector with dates to attribute to the rows of the dataframe.}
}
\value{
A new dataframe , obtained by eliminating the first column and denominating rows with the dates we chose.
}
\description{
The function retrieves the data we import, attributing a date vector to the rows.
}
\examples{
tab <- frmtb(data,c('30/06/2016','30/06/2017','30/06/2018'))
}
