% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R PACKAGE (DEFINITIVO).R
\name{grph}
\alias{grph}
\title{Plotly Bars Graph Generator}
\usage{
grph(z)
}
\arguments{
\item{z}{A vector with yearly Break-even values.}
}
\value{
An interactive bars graph with 3 parameters of interest.
}
\description{
The function generates an interactive plotly bars graph. On the x-bar we put the rownames as dates of the dataframe of our interest. On the first y-bar we insert the aggregate values of Relevant Expenses ordered by year. On the second y-bar we insert the aggregate values of Relevant Revenues by year. On the third y-bar we insert the vector we obtain by running the beyR() function.
}
