% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R PACKAGE (DEFINITIVO).R
\name{rel_var}
\alias{rel_var}
\title{Default Relevant Variables Function}
\usage{
rel_var(df, vct_var)
}
\arguments{
\item{df}{A dataframe of our interest.}

\item{vct_var}{The vector with (ordered) columns we want to extract from the original dataframe.}
}
\value{
A new dataframe with the columns we assigned to the vct_var.
}
\description{
The function creates a new dataframe with selected columns retrived by vct_var.
}
\examples{
RR <-rel_var(tab,c(2,3,5,6,7) RE <- rel_var(tab,c(10,12,13,14,15))
}
