% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R PACKAGE (DEFINITIVO).R
\name{shinyapp_asr}
\alias{shinyapp_asr}
\title{Relevant Accounts ShinyAppAsr}
\usage{
shinyapp_asr(server, ui)
}
\arguments{
\item{server}{Default parameter by shinyApp interface}

\item{ui}{Default parameter by shinyApp interface}
}
\value{
An interactive page with bar graphs and selection of relevant accounts through a sidebar panel.
}
\description{
Relevant Accounts ShinyAppAsr
}
