% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R PACKAGE (DEFINITIVO).R
\name{sum_be}
\alias{sum_be}
\title{Break-Even Calculator}
\usage{
sum_be(df1, df2)
}
\arguments{
\item{df1}{The dataframe we obtained from rel_vaRR(), including the column obtained by running the trnsfR() function.}

\item{df2}{The dataframe we obtained from rel_vaRE(), including the column obtained by running the trnsfR() function.}
}
\value{
A numeric value, indicating the break-even value for the 3-years monitoring period under inspection.
}
\description{
The function print the result of the sum between Aggregate Revenues and Aggregate Expenses (i.e. "break-even requirement" under the UEFA Financial Fair Play Regulation 2019), computed for the 3-years monitoring period.
}
\examples{
be <- sum_be(RR,RE)
}
