% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R PACKAGE (DEFINITIVO).R
\name{trnsfR}
\alias{trnsfR}
\title{Columns Transfer Function}
\usage{
trnsfR(df)
}
\arguments{
\item{df}{The dataframe we obtained from rel_vaRR() and rel_vaRE().}
}
\value{
A new columns to be add to RR and RE whose values are retrived as the aggregate rows sum of all of the columns.
}
\description{
The function adds columns with row sums to our dataframes of interest.
}
\examples{
RR$AggrRev <- trnsfR(RR) RE$AggrExp <- trnsfR(RE)
}
