% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R PACKAGE (DEFINITIVO).R
\name{upload}
\alias{upload}
\title{Upload Function}
\usage{
upload()
}
\value{
As csv file, the data we want to use for peculiar package purposes.
}
\description{
The function upload a file suitable for the package in a csv format.
}
\examples{
DATA <- upload()
}
